editor/libeditor/tests/test_undo_redo_stack_after_setting_value.html

<!DOCTYPE html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1473515
-->
<html>
<head>
  <title>Test for Bug 1473515</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1473515">Mozilla Bug 1473515</a>
<p id="display"></p>
<div id="content" style="display: none;">

</div>

<input id="input">
<textarea id="textarea"></textarea>

<pre id="test">

<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  let editableElements = [
    document.getElementById("input"),
    document.getElementById("textarea"),
  ];
  for (let editableElement of editableElements) {
    function checkInputEvent(aEvent, aInputType, aData, aDescription) {
      ok(aEvent instanceof InputEvent,
         `"input" event should be dispatched with InputEvent interface ${aDescription}`);
      is(aEvent.cancelable, false,
         `"input" event should be never cancelable ${aDescription}`);
      is(aEvent.bubbles, true,
         `"input" event should always bubble ${aDescription}`);
      is(aEvent.inputType, aInputType,
         `inputType should be "${aInputType}" ${aDescription}`);
      is(aEvent.data, aData,
         `data should be ${aData} ${aDescription}`);
      is(aEvent.dataTransfer, null,
         `dataTransfer should be null ${aDescription}`);
    }

    let inputEvents = [];
    function onInput(aEvent) {
      inputEvents.push(aEvent);
    }
    editableElement.addEventListener("input", onInput);

    editableElement.focus();

    inputEvents = [];
    synthesizeKey("a");
    is(inputEvents.length, 1,
       `Only one "input" event should be fired when inserting "a" with key on <${editableElement.tagName.toLowerCase()}> element`);
    checkInputEvent(inputEvents[0], "insertText", "a", `when inserting "a" with key on <${editableElement.tagName.toLowerCase()}> element`);

    inputEvents = [];
    synthesizeKey("c");
    is(inputEvents.length, 1,
       `Only one "input" event should be fired when inserting "c" with key on <${editableElement.tagName.toLowerCase()}> element`);
    checkInputEvent(inputEvents[0], "insertText", "c", `when inserting "c" with key on <${editableElement.tagName.toLowerCase()}> element`);

    inputEvents = [];
    synthesizeKey("KEY_ArrowLeft");
    is(inputEvents.length, 0,
       `No "input" event should be fired when pressing "ArrowLeft" key on <${editableElement.tagName.toLowerCase()}> element`);

    inputEvents = [];
    synthesizeKey("b");
    is(inputEvents.length, 1,
       `Only one "input" event should be fired when inserting "b" with key on <${editableElement.tagName.toLowerCase()}> element`);
    checkInputEvent(inputEvents[0], "insertText", "b", `when inserting "b" with key on <${editableElement.tagName.toLowerCase()}> element`);

    let editor = SpecialPowers.wrap(editableElement).editor;
    let transactionManager = editor.transactionManager;
    is(transactionManager.numberOfUndoItems, 2,
       editableElement.tagName + ": Initially, there should be 2 undo items");
    // Defined as nsITextControlElement::DEFAULT_UNDO_CAP
    is(transactionManager.maxTransactionCount, 1000,
       editableElement.tagName + ": Initially, transaction manager should be able to have 1,000 undo items");

    inputEvents = [];
    editableElement.value = "def";
    is(inputEvents.length, 0,
       `No "input" event should be fired when setting value of <${editableElement.tagName.toLowerCase()}> element`);

    is(transactionManager.numberOfUndoItems, 0,
       editableElement.tagName + ": After setting value, all undo items must be deleted");
    is(transactionManager.maxTransactionCount, 1000,
       editableElement.tagName + ": After setting value, maximum transaction count should be restored to the previous value");

    inputEvents = [];
    synthesizeKey("a");
    is(inputEvents.length, 1,
       `Only one "input" event should be fired when inserting "a" with key again on <${editableElement.tagName.toLowerCase()}> element`);
    checkInputEvent(inputEvents[0], "insertText", "a", `when inserting "a" with key again on <${editableElement.tagName.toLowerCase()}> element`);

    inputEvents = [];
    synthesizeKey("z", { accelKey: true });
    is(editableElement.value, "def",
       editableElement.tagName + ": undo should work after setting value");
    is(inputEvents.length, 1,
       `Only one "input" event should be fired when undoing on <${editableElement.tagName.toLowerCase()}> element`);
    checkInputEvent(inputEvents[0], "historyUndo", null, `when undoing on <${editableElement.tagName.toLowerCase()}> element`);

    // Disable undo/redo.
    editor.enableUndo(0);
    is(transactionManager.maxTransactionCount, 0,
       editableElement.tagName + ": Transaction manager should not be able to have undo items");
    editableElement.value = "hij";
    is(transactionManager.maxTransactionCount, 0,
       editableElement.tagName + ": Transaction manager should not be able to have undo items after setting value");

    editableElement.removeEventListener("input", onInput);
  }
  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
